Filter out duplicate alerts for a resource/rule
In order to deduplicate a stream of arbritrary alerts we will want to
know the rule that generated them, the resource that cause it and the
message. This code adds the associated rule, which was the only missing
information.

I had to restructure the code a little in order to allow an alert to
reference a rule which was in the same file that referenced alerts. Yay
cycles!